youtube-curiosity/requirements.txt
accelerate==0.26.1
aiohttp==3.9.1
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
audioread==3.0.1
boto3==1.34.45
botocore==1.34.45
Brotli==1.1.0
cachetools==5.3.3
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
clip @ git+https://github.com/openai/CLIP.git@a1d071733d7111c9c014f024669f959182114e33
datasets==2.16.1
decorator==4.4.2
dill==0.3.7
filelock==3.13.1
frozenlist==1.4.1
fsspec==2023.10.0
ftfy==6.1.3
google-api-core==2.17.1
google-api-python-client==2.119.0
google-auth==2.28.1
google-auth-httplib2==0.2.0
googleapis-common-protos==1.62.0
httplib2==0.22.0
huggingface-hub==0.20.2
idna==3.6
imageio==2.33.1
imageio-ffmpeg==0.4.9
importlib-metadata==7.0.1
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.3.2
lazy_loader==0.3
librosa==0.10.1
llvmlite==0.41.1
MarkupSafe==2.1.3
# mkl-fft @ file:///croot/mkl_fft_1695058164594/work
# mkl-random @ file:///croot/mkl_random_1695059800811/work
# mkl-service==2.4.0
mpmath==1.3.0
msgpack==1.0.7
multidict==6.0.4
multiprocess==0.70.15
mutagen==1.47.0
networkx==3.1
nltk==3.8.1
numba==0.58.1
# numpy @ file:///work/mkl/numpy_and_numpy_base_1682953417311/work
numpy
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
opencv-python==4.9.0.80
packaging==23.2
pandas==2.0.3
Pillow==9.3.0
platformdirs==4.1.0
pooch==1.8.0
proglog==0.1.10
protobuf==4.25.3
psutil==5.9.7
pyarrow==14.0.2
pyarrow-hotfix==0.6
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodomex==3.20.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.1
pytube @ git+https://github.com/24makee/pytube.git@c709202d4f2c0d36d9484314d44fd26744225b7d
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
rsa==4.9
s3transfer==0.10.0
safetensors==0.4.1
scikit-learn==1.3.2
scipy==1.10.1
six==1.16.0
soundfile==0.12.1
soxr==0.3.7
sympy==1.12
threadpoolctl==3.2.0
tokenizers==0.15.0
torch==2.1.2
torchvision==0.8.2
tqdm==4.66.1
transformers==4.36.2
triton==2.1.0
# typing_extensions @ file:///croot/typing_extensions_1705005625920/work
tzdata==2023.4
uritemplate==4.1.1
urllib3==1.26.18
wcwidth==0.2.13
websockets==12.0
xxhash==3.4.1
yarl==1.9.4
yt-dlp==2023.12.30
zipp==3.17.0
